Add repo digest to workloadmeta.container from docker runtime collection (#23045)
Add repo digest to workloadmeta.container from docker runtime collection
