+++ comp/core/workloadmeta/collectors/internal/containerd/container_builder.go
@@ -20,6 +20,7 @@ import (
 	"github.com/containerd/containerd/namespaces"
 
 	"github.com/DataDog/datadog-agent/comp/core/workloadmeta"
+	"github.com/DataDog/datadog-agent/comp/core/workloadmeta/collectors/util"
 	cutil "github.com/DataDog/datadog-agent/pkg/util/containerd"
 	"github.com/DataDog/datadog-agent/pkg/util/containers"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
@@ -27,11 +28,6 @@ import (
 
 // kataRuntimePrefix is the prefix used by Kata Containers runtime
 const kataRuntimePrefix = "io.containerd.kata"
-
-const (
-	// SHA256 is the prefix used by containerd for the repo digest
-	SHA256 = "@sha256:"
-)
 
 // buildWorkloadMetaContainer generates a workloadmeta.Container from a containerd.Container
 func buildWorkloadMetaContainer(namespace string, container containerd.Container, containerdClient cutil.ContainerdItf, store workloadmeta.Component) (workloadmeta.Container, error) {
@@ -66,7 +62,7 @@ func buildWorkloadMetaContainer(namespace string, container containerd.Container
 		log.Debugf("cannot split image name %q: %s", info.Image, err)
 	}
 
-	image.RepoDigest = extractRepoDigestFromImage(imageID, image.Registry, store) // "sha256:digest"
+	image.RepoDigest = util.ExtractRepoDigestFromImage(imageID, image.Registry, store) // "sha256:digest"
 
 	status, err := containerdClient.Status(namespace, container)
 	if err != nil {
@@ -160,57 +156,4 @@ func extractRuntimeFlavor(runtime string) workloadmeta.ContainerRuntimeFlavor {
 		return workloadmeta.ContainerRuntimeFlavorKata
 	}
 	return workloadmeta.ContainerRuntimeFlavorDefault
-}
-
-// extractRepoDigestFromImage extracts the repoDigest from workloadmeta store if it exists and unique
-// the format of repoDigest is "registry/repo@sha256:digest", the format of return value is "sha256:digest"
-func extractRepoDigestFromImage(imageID string, imageRegistry string, store workloadmeta.Component) string {
-	existingImgMetadata, err := store.GetImage(imageID)
-	if err == nil {
-		// If there is only one repoDigest, return it
-		if len(existingImgMetadata.RepoDigests) == 1 {
-			_, _, digest := parseRepoDigest(existingImgMetadata.RepoDigests[0])
-			return digest
-		}
-		// If there are multiple repoDigests, we should find the one that matches the current container's registry
-		for _, repoDigest := range existingImgMetadata.RepoDigests {
-			registry, _, digest := parseRepoDigest(repoDigest)
-			if registry == imageRegistry {
-				return digest
-			}
-		}
-		log.Debugf("cannot get matched registry in repodigests for image %s", imageID)
-	} else {
-		// TODO: we should handle the case when the image metadata is not found in the store
-		// For example, there could be a rare race condition when collection of image metadata is not finished yet
-		// In this case, we can either call containerd client directly or get it from knownImages in the local cache
-		// In fact, knownImages is already updated in the same goroutine, so it should be available
-		// If it is not available, containerd client can be used to get the image metadata
-		log.Debugf("cannot get image metadata for image %s: %s", imageID, err)
-	}
-	return ""
-}
-
-// parseRepoDigest extracts registry, repository, digest from repoDigest (in the format of "registry/repo@sha256:digest")
-func parseRepoDigest(repoDigest string) (string, string, string) {
-	var registry, repository, digest string
-	imageName := repoDigest
-
-	digestStart := strings.Index(repoDigest, SHA256)
-	if digestStart != -1 {
-		digest = repoDigest[digestStart+len("@"):]
-		imageName = repoDigest[:digestStart]
-	}
-
-	registryEnd := strings.Index(imageName, "/")
-	// e.g <registry>/<repository>
-	if registryEnd != -1 {
-		registry = imageName[:registryEnd]
-		repository = imageName[registryEnd+len("/"):]
-		// e.g <registry>
-	} else {
-		registry = imageName
-	}
-
-	return registry, repository, digest
 }
